+++ src/modules/macros/macro-builder.ts
@@ -1,0 +1,15 @@
+import jsonQuery from "json-query"
+
+abstract class MacroBuilder {
+    steps: []
+
+    constructor() {
+
+    }
+
+    addStep() { }
+
+    create(): Macro {
+        return null
+    }
+}