+++ CHANGELOG.md
@@ -1,10 +1,20 @@
+## [1.3]
+
+### Added
+ * Beta custom chat log and option to revert to vanilla log. This introduces the ability to keep timers in the chat log and sort by IC/OOC messages
+ * All rolls from the sheet will be aliased
+
+### Fixed
+ * Bug where disadvantages and advantages would swap places
+ * Bug where CR was available for advantages but not for disadvantages 
+
 ## [1.2.2]
 
-## Added
+### Added
 * changelog section
 * project info section
 * quick roll against disadvantage control rating
 
-## Fixed
+### Fixed
 * Bug that prevented the removal of equipment
 * GCS Importer